requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
python-jose[cryptography]
pwdlib[argon2]
python-multipart
pytest
httpx
email-validator